move test_set and test_get back to helpers
