+++ unittest/gunit/villagesql/type_context-t.cc
@@ -23,7 +23,6 @@
 
 #include "unittest/gunit/test_utils.h"
 #include "villagesql/schema/descriptor/type_context.h"
-#include "villagesql/schema/identifier_names.h"
 #include "villagesql/schema/systable/helpers.h"
 #include "villagesql/sdk/include/villagesql/abi/types.h"
