+++ villagesql/schema/systable/helpers.cc
@@ -206,4 +206,11 @@ std::string normalize_index_name(const std::string &name) {
   return casedn(&my_charset_utf8mb4_0900_ai_ci, name);
 }
 
+// ===== Test utilities =====
+void test_set_lower_case_table_names(int value) {
+  ::lower_case_table_names = value;
+}
+
+int test_get_lower_case_table_names() { return ::lower_case_table_names; }
+
 }  // namespace villagesql
